more clean fix

+++ build-opt/build_5.x_pgo.sh
@@ -2,7 +2,11 @@
 # Created by Roel Van de Paar, Percona LLC
 # Patched by Ben(bash99@gmail.com) for automatic pgo
 
+NPROC=$(nproc)
+MJ=$(($NPROC*3/2))
 INSPATH=/usr
+export MYSQL_USER=`whoami`
+export MYSQL_BASE=`pwd`/local/mysql
 
 cd ps-5.7
 
@@ -31,6 +35,7 @@ cp -al ${CURPATH} ${CURPATH}_opt
 cd ${CURPATH}_opt
 
 patch -p1 < ../pspgo-utils/build-opt/cmake-pgo.patch
+if [ $? -ne 0 ]; then echo "patch failed! Assert: non-0 exit status detected!"; exit 1; fi
 
 ### TEMPORARY HACK TO AVOID COMPILING TB (WHICH IS NOT READY YET)
 rm -Rf ./plugin/tokudb-backup-plugin
@@ -38,16 +43,10 @@ rm -Rf ./plugin/tokudb-backup-plugin
 cmake . -DWITH_ZLIB=system -DBUILD_CONFIG=mysql_release -DFEATURE_SET=community -DWITH_NUMA=ON -DWITH_SYSTEMD=1 -DWITH_EMBEDDED_SERVER=OFF -DWITH_TOKUDB=0 -DWITH_ROCKSDB=0 -DDOWNLOAD_BOOST=0 -DWITH_BOOST=../boost_1_59_0 -DWITH_SSL=bundled -DWITH_MECAB=/opt/rh/rh-mysql57/root/usr/ -DENABLE_DOWNLOADS=1 -DWITH_PAM=ON | tee /tmp/5.7_opt_build
 if [ $? -ne 0 ]; then echo "Assert: non-0 exit status detected!"; exit 1; fi
   
-make -j12 | tee -a /tmp/5.7_opt_build
-if [ $? -ne 0 ]; then echo "Assert: non-0 exit status detected!"; exit 1; fi
+make -j$MJ | tee -a /tmp/5.7_opt_build
+if [ $? -ne 0 ]; then echo "compile pgo gene failed! Assert: non-0 exit status detected!"; exit 1; fi
 
-#chown -R mysql:mysql .
-
-cd ..
-export MYSQL_USER=`whoami`
-export MYSQL_BASE=`pwd`/local/mysql
-
-LD_PRELOAD="/usr/lib64/libjemalloc.so.1" MALLOC_CONF="lg_dirty_mult:-1" ${CURPATH}_opt/sql/mysqld --defaults-file=$MYSQL_BASE/etc/my.automem.cnf --basedir=$MYSQL_BASE --datadir=$MYSQL_BASE/data --plugin-dir=$MYSQL_BASE/lib/mysql/plugin --user=$MYSQL_USER &
+LD_PRELOAD="/usr/lib64/libjemalloc.so.1" MALLOC_CONF="lg_dirty_mult:-1" sql/mysqld --defaults-file=$MYSQL_BASE/etc/my.automem.cnf --basedir=$MYSQL_BASE --datadir=$MYSQL_BASE/data --plugin-dir=$MYSQL_BASE/lib/mysql/plugin --user=$MYSQL_USER &
 #### waiting buffer pool load
 sleep 60
 
@@ -60,9 +59,10 @@ $MYSQL_BASE/bin/mysqladmin -u root --socket=$MYSQL_BASE/data/mysql.sock shutdown
 find . -name "flags.make" | xargs -n 64 perl -pi -e "s/profile-generate/profile-use -fprofile-correction /g"
 find . -name "link.txt" | xargs -n 64 perl -pi -e "s/profile-generate/profile-use -fprofile-correction /g"
 
-make -j12
+make -j$MJ
+if [ $? -ne 0 ]; then echo "compile pgo-use failed Assert: non-0 exit status detected!"; exit 1; fi
 
-LD_PRELOAD="/usr/lib64/libjemalloc.so.1" MALLOC_CONF="lg_dirty_mult:-1" ${CURPATH}_opt/sql/mysqld --defaults-file=$MYSQL_BASE/etc/my.automem.cnf --basedir=$MYSQL_BASE --datadir=$MYSQL_BASE/data --plugin-dir=$MYSQL_BASE/lib/mysql/plugin --user=$MYSQL_USER &
+LD_PRELOAD="/usr/lib64/libjemalloc.so.1" MALLOC_CONF="lg_dirty_mult:-1" sql/mysqld --defaults-file=$MYSQL_BASE/etc/my.automem.cnf --basedir=$MYSQL_BASE --datadir=$MYSQL_BASE/data --plugin-dir=$MYSQL_BASE/lib/mysql/plugin --user=$MYSQL_USER &
 #### waiting buffer pool load
 sleep 60
 
@@ -71,3 +71,4 @@ sh ../pspgo-utils/build-opt/train-sysbench.sh | tee /tmp/train_result.txt
 
 $MYSQL_BASE/bin/mysqladmin -u root --socket=$MYSQL_BASE/data/mysql.sock shutdown
 
+grep transaction /tmp/*_result.txt | tee ../pgo_result.txt
